pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "virtual-thermostat"
version = "0.1.0"
description = "A virtual thermostat for controlling smart plugs based on temperature"
authors = [
    {name = "Vicente Bolea"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "python-kasa",
    "trame",
    "trame-vuetify",
    "trame-server",
]

[project.scripts]
vthermostat-cli = "virtual_thermostat.cli:cli_main"
vthermostat-ui = "virtual_thermostat.ui:main"
vthermostat-daemon = "virtual_thermostat.daemon:main"
vthermostat-dht11 = "virtual_thermostat.dht11:main"

[tool.setuptools]
packages = ["virtual_thermostat"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
include = 'virtual_thermostat/.*\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "flake8",
    "pytest",
]
hardware = [
    "Adafruit-DHT",
]
